Remove icon sets and Tina editor runtime from client bundles
- Lazy-load the full react-icons map behind dynamicIcon (site Icon renders
  a same-size placeholder until the async chunk loads)
- Replace named react-icons/si + /bi imports in render components with
  local inline SVGs (named imports link the whole multi-MB collection)
- Split Tina schemas out of component files (tableLayout, iconLabel) and
  move paymentDetailsBlockConstant to a dependency-free file so admin-only
  code (wrapFieldsWithMeta, pickers) is only reachable from tina/config
- Track /events/ai-workshop-essentials in the Lighthouse workflow
