+++ .github/workflows/template-lighthouse.yml
@@ -52,6 +52,7 @@ jobs:
             ${{ inputs.url }}/employment
             ${{ inputs.url }}/events/angular-superpowers-tour
             ${{ inputs.url }}/events/ai-workshop
+            ${{ inputs.url }}/events/ai-workshop-essentials
             ${{ inputs.url }}/netug/sydney
             ${{ inputs.url }}/training/internship-fullstack
