+++ components/blocks/payment-block.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import type { Template } from "tinacms";
 import { tinaField } from "tinacms/dist/react";
 import { TinaMarkdown } from "tinacms/dist/rich-text";
-import { paymentDetailsBlockConstant } from "../../tina/collections/payment-details"; // TODO: Use alias - https://github.com/tinacms/tinacms/issues/4488
+import { paymentDetailsBlockConstant } from "../../tina/collections/payment-details.constant";
 
 export type PaymentBlockProps = {
   bankName: string;
