+++ components/blocksSubtemplates/tinaFormElements/colourOptions/buttonOptions.tsx
@@ -1,4 +1,4 @@
-import { ColorPickerOptions } from "../colourSelector";
+import type { ColorPickerOptions } from "../colourSelector";
 
 export const buttonOptions: ColorPickerOptions[] = [
   {
